Move Cim2NdlMapper from router.model to network.capability.basic.
Dependency is inverted and router.model knows nothing about networks.
Furthermore, network.model remains unaware of router.model.

Translation from router model to network model is done in network.capability.basic,
so this capability has dependencies to both models.
There are other network capabilities with references to both models,
if they make use of router.model to make calls to router capabilities.

This patch moves the mapper and updates dependencies in features as well as pom files.
Integration tests pass with clean m2 repo.
