+++ core/features/src/main/resources/features.xml
@@ -36,10 +36,6 @@
 
 	<feature name="opennaas-cim" version="${project.version}">
 		<feature version="${opennaas.version}">org.opennaas.core</feature>
-
-		<!-- The core model unfortunately depends on the network model
-		     - it should be the other way around. -->
-		<bundle>mvn:org.opennaas/org.opennaas.extensions.network.model/${project.version}</bundle>
 		<bundle>mvn:org.opennaas/org.opennaas.extensions.router.model/${project.version}</bundle>
 
 		<!-- QueueManager should probably be part of org.opennaas.core,
@@ -115,11 +111,13 @@
 	<feature name="opennaas-network" version="${project.version}">
 		<feature version="${project.version}">opennaas-cim</feature>
 
+		<bundle>mvn:org.opennaas/org.opennaas.extensions.network.model/${project.version}</bundle>
 		<bundle>mvn:org.opennaas/org.opennaas.extensions.network.repository/${project.version}</bundle>
 		<bundle>mvn:org.opennaas/org.opennaas.extensions.network.capability.basic/${project.version}</bundle>
 		<bundle>mvn:org.opennaas/org.opennaas.extensions.network.capability.queue/${project.version}</bundle>
 		<bundle>mvn:org.opennaas/org.opennaas.extensions.network.capability.ospf/${project.version}</bundle>
-		<bundle>mvn:org.opennaas/org.opennaas.extensions.router.capability.ospf/${project.version}</bundle>
+		<bundle dependency="true">mvn:org.opennaas/org.opennaas.extensions.router.capability.ospf/${project.version}</bundle>
+		
 	</feature>
 
 	<feature name="opennaas-bod" version="${project.version}">
@@ -127,6 +125,7 @@
 
 		<bundle dependency="true">mvn:joda-time/joda-time/${jodatime.version}</bundle>
 		<bundle dependency="true">mvn:com.googlecode.guava-osgi/guava-osgi/${guava.version}</bundle>
+		<bundle dependency="true">mvn:org.opennaas/org.opennaas.extensions.network.model/${project.version}</bundle>
 
 		<bundle>mvn:org.opennaas/org.opennaas.extensions.bod.repository/${project.version}</bundle>
 		<bundle>mvn:org.opennaas/org.opennaas.extensions.bod.capability.l2bod/${project.version}</bundle>
