+++ extensions/bundles/bod.repository/pom.xml
@@ -20,7 +20,7 @@
 		</dependency>
 		<dependency>
 			<groupId>org.opennaas</groupId>
-			<artifactId>org.opennaas.extensions.router.model</artifactId>
+			<artifactId>org.opennaas.extensions.network.model</artifactId>
 		</dependency>
 		<dependency>
 			<groupId>org.opennaas</groupId>
