+++ extensions/bundles/network.capability.basic/pom.xml
@@ -22,12 +22,11 @@
 			<groupId>org.opennaas</groupId>
 			<artifactId>org.opennaas.extensions.network.model</artifactId>
 		</dependency>
-		<!-- FIXME remove when Model2NdlMapper is in IModel instead of mantychore.model -->
 		<dependency>
 			<groupId>org.opennaas</groupId>
 			<artifactId>org.opennaas.extensions.router.model</artifactId>
 		</dependency>
-		<!-- FIXME capability should not depend on repo, mapper should be moved-->
+		<!-- FIXME capability should not depend on repo, ModelToDescriptorMapper should be moved-->
 		<dependency>
 			<groupId>org.opennaas</groupId>
 			<artifactId>org.opennaas.extensions.network.repository</artifactId>
