+++ extensions/bundles/network.capability.ospf/pom.xml
@@ -18,6 +18,10 @@
 			<groupId>org.opennaas</groupId>
 			<artifactId>org.opennaas.extensions.router.capability.ospf</artifactId>
 		</dependency>
+		<dependency>
+			<groupId>org.opennaas</groupId>
+			<artifactId>org.opennaas.extensions.network.model</artifactId>
+		</dependency>
 	</dependencies>
 
 	<build>
