+++ extensions/bundles/network.capability.queue/pom.xml
@@ -26,6 +26,10 @@
 			<groupId>org.opennaas</groupId>
 			<artifactId>org.opennaas.extensions.queuemanager</artifactId>
 		</dependency>
+		<dependency>
+			<groupId>org.opennaas</groupId>
+			<artifactId>org.opennaas.extensions.network.model</artifactId>
+		</dependency>
 	</dependencies>
 	<build>
 		<plugins>
