+++ itests/bod/pom.xml
@@ -20,6 +20,10 @@
 			<groupId>org.opennaas</groupId>
   			<artifactId>org.opennaas.extensions.bod.capability.l2bod</artifactId>
 		</dependency>
+		<dependency>
+			<groupId>org.opennaas</groupId>
+			<artifactId>org.opennaas.extensions.network.model</artifactId>
+		</dependency>
 		
 		<!--  FIXME check if necessary and delete otherwise -->
 		<dependency>
@@ -38,6 +42,7 @@
 			<groupId>org.opennaas.itests</groupId>
 			<artifactId>org.opennaas.itests.helpers</artifactId>
 		</dependency>
+		
 		
 	</dependencies>
